move strategies into their own file
